issue-9: first pass at reorg
