+++ src/afwerx_datathon/learning/__init__.py
@@ -1,0 +1,1 @@
+"""Modules for machine learning."""
