+++ src/afwerx_datathon/learning/supervised/__init__.py
@@ -1,0 +1,1 @@
+"""Modules for supervised learning."""
