+++ src/afwerx_datathon/signal/__init__.py
@@ -1,0 +1,1 @@
+"""Modules for signal analysis and DSP."""
